refactor: selectors
